bug fix and save patients to json file
